graph_ml/requirements.txt
tensorflow-macos~=2.9.2
numpy~=1.22.4